autoreload compatibility
This is related to this issue:

https://stackoverflow.com/questions/32481508/autoreload-and-package-causing-typeerror-supertype-obj-obj-must-be-an-insta

Autoreload was throwing an error when editing code and re-running a notebook cell.

This syntax for calling the parent method is simpler, legit, and solved that problem for me.

I used the regex ``\([A-Z]([^)]+), self\)`` to find all the matches in code, so this would be a comprehensive update were we to merge that change.