Fixes #407 - log(0) warnings
When the log-library size is calculated, the dataset may contain empty
cells, which will then trigger `log(0)' warnings.
We do not want these unintelligible warnings, nor do we want the data to
be filtered without the user's permission.

In this commit, we capture these warnings using masked numpy array and
warn the user with a more explicit message