Fix reinitialization error in GIMVI (#2446)
Fixes #2273. Error occurs when re-initializing GIMVI with the same
spatial dataset due to the following code in `GIMVI.__init__` that
in-place modified an `.obs` column:
```
adata_seq_n_batches = sum_stats[0]["n_batch"]
adata_spatial.obs[
    self.adata_managers["spatial"]
    .data_registry[REGISTRY_KEYS.BATCH_KEY]
    .attr_key
] += adata_seq_n_batches
```
This was necessary as the spatial batches and GEX batches were
0-indexed, but we wanted to distinguish between them. The fix checks
whether the obs column's minimum is 0, and if so, we add the offset.
Otherwise, we don't, since that means the column was updated before
within the same session.