Add API for custom dataloaders in `SCVI` (#2467)
This commit enables users to bypass the default workflow of setting up
an `AnnData` and initializing a model with it in order to train, and
instead being able to train a model directly with a custom dataloader.

- Make `adata` argument optional in `SCVI`, which sets a
`_module_init_on_train` flag (through `BaseModelClass`) if no `adata` is
passed in
- This causes the module init to be delayed until train time since we
won't have access to certain arguments (*e.g.* `n_input` and `n_batch`)
that must be inferred from the data
- Add an optional `data_module` argument to
`UnsupervisedTrainingMixin.train` that must be a `LightningDataModule`
and must be passed in if the model was not initialized with `adata`

Set as an experimental feature for now. Will expand to models other than
`SCVI` when the implementation and API stabilize.