feat(model): support custom dataloaders in vaemixin (#2748)
adds custom dataloader arguments to vaemixin methods. currently does not
check that the dataloader outputs in the correct format such that errors
will be raised at the module level, which will potentially be confusing.

also includes a refactor of some of the vaemixin methods, including more
detailed and consistent docstrings.