feat(model): save mudata var names per modality (#2769)
previously, mudata-based models (totalvi and tangram) saved variable
names in a flat array with values concatenated across all modalities -
this is what happens when MuData.var_names is called directly. this is
an issue for performing scarches on such a model as we cannot pad with
zeros without being able to map individual variable names back to their
respective modalities.

this commit sets the default to saving per-modality var names as a
dictionary, with keys corresponding to modality keys and values to
variable names. additionally, it adds a flag to the save method in case
a user wants to use the legacy saving format.

old model saves will still be loaded successfully as the code will
detect if the legacy save was used and validate var_names accordingly.

BREAKING CHANGE: changes the default saving scheme for mudata-based
models. see above for details.