tutorial index page with filterable list (#3276)
# Custom sphinx directives for filterable lists
Created custom sphinx directives which can be used to create a
filterable list of cards, and used these directives to create a
filterable list of tutorial cards for the tutorials index page.

These directives could also be used to create similar filterable lists
easily if needed in other parts of the documentation.

## customcard
The first custom directive is `customcard` which is used to create, in
this case, a card displaying information about a single scvi-tools
tutorial. It has `path` parameter for specifying the path in the
documentation to the tutorial, and a `tags` parameter for the tags
describing that tutorial, which will be used for filtering in the list.
The description for the tutorial goes in the directive's content.

This directive automatically generates the title, description, and group
name that will be displayed on the card, as long as the correct path has
been provided. The `customcard` directive should be placed a model
group's index file (such as index_scrna.md).

Example:
```
\```{customcard}
:path: notebooks/scrna/harmonization
:tags: Analysis, Integration, Transfer-learning, Dimensionality-reduction, Removal-of-variance

This is a test description. Put the description as the main content of the customcard directive.
\```
```
## customlist
The second custom directive is `customlist` which creates a filterable
list from all of the customcard directives. During the documentation
build process, this list will be generated in the markdown file where
`customlist` is placed, such as index.md. No parameters are needed for
this directive and it automatically retrieves all of the cards placed
throughout the documentation files and adds them. It also generates
buttons for filtering tutorials based on all found tags, and generates
tabs for selecting the tutorial group to filter by (scRNA-seq,
scATAC-seq, Quick Start...).

Example:
```
\```{customlist}
\```
```

## Tutorial List
I've used these directives to create a list for the tutorials, which can
be seen in the docs build preview on the tutorials page.

### Order of tutorials
First of all, the tutorial group order is specified inside of the file
called filterlist.py, in a list at the beginning of the file. Examples
of tutorial groups are scRNA, Quickstart, etc. If a new group is added,
then the group name should be added to this list. Within each group, the
order is determined by the order of the customcard directives within
each group's index file.